// packages/datadog-api-client-v2/models/LogsWarning.d.ts
/**
 * Unless explicitly stated otherwise all files in this repository are licensed under the Apache-2.0 License.
 * This product includes software developed at Datadog (https://www.datadoghq.com/).
 * Copyright 2020-Present Datadog, Inc.
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
/**
* A warning message indicating something that went wrong with the query
*/
export declare class LogsWarning {
    /**
    * A unique code for this type of warning
    */
    'code'?: string;
    /**
    * A detailed explanation of this specific warning
    */
    'detail'?: string;
    /**
    * A short human-readable summary of the warning
    */
    'title'?: string;
    static readonly discriminator: string | undefined;
    static readonly attributeTypeMap: {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static getAttributeTypeMap(): {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static deserialize(data: {
        [key: string]: any;
    }): LogsWarning;
    static serialize(data: LogsWarning): {
        [key: string]: any;
    };
    constructor();
}